requirements.txt
#RUN sudo pip install -r requirements.txt TO INSTALL ALL THE PACKAGE REQUIREMENTS
beautifulsoup4==4.11.2
cryptography==40.0.2
googletrans==3.1.0a0
gTTS==2.3.1
keyboard==0.13.5
nltk==3.7
numpy==1.24.2
openai==0.26.5
outcome==1.2.0
pandas==2.0.0
Pillow==9.4.0
pip==23.0.1
playsound==1.3.0
pvporcupine==2.2.0
pyaudio==0.2.13
pyautogui==0.9.53
pycparser==2.21
pycryptodome==3.17
pyttsx3==2.90
pywhatkit==5.4
pywin32==305
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
selenium==4.1.3
setuptools==67.6.1
speechrecognition==3.9.0
torch2.0.0
wikipedia==1.4.0
websockets==10.4
wolframalpha==5.0.0
